+++ libc/src/string/CMakeLists.txt
@@ -35,24 +35,6 @@ add_header_library(
 )
 
 add_entrypoint_object(
-  bcopy
-  SRCS
-    bcopy.cpp
-  HDRS
-    bcopy.h
-)
-
-add_entrypoint_object(
-  index
-  SRCS
-    index.cpp
-  HDRS
-    index.h
-  DEPENDS
-    .string_utils
-)
-
-add_entrypoint_object(
   memccpy
   SRCS
     memccpy.cpp
@@ -96,16 +78,6 @@ add_entrypoint_object(
     memrchr.cpp
   HDRS
     memrchr.h
-)
-
-add_entrypoint_object(
-  rindex
-  SRCS
-    rindex.cpp
-  HDRS
-    rindex.h
-  DEPENDS
-    .string_utils
 )
 
 add_entrypoint_object(
@@ -169,17 +141,6 @@ add_entrypoint_object(
     strcmp.h
   DEPENDS
     .memory_utils.inline_strcmp
-)
-
-add_entrypoint_object(
-  strcasecmp
-  SRCS
-    strcasecmp.cpp
-  HDRS
-    strcasecmp.h
-  DEPENDS
-    .memory_utils.inline_strcmp
-    libc.src.__support.ctype_utils
 )
 
 add_entrypoint_object(
@@ -317,17 +278,6 @@ add_entrypoint_object(
     strncmp.h
   DEPENDS
     .memory_utils.inline_strcmp
-)
-
-add_entrypoint_object(
-  strncasecmp
-  SRCS
-    strncasecmp.cpp
-  HDRS
-    strncasecmp.h
-  DEPENDS
-    .memory_utils.inline_strcmp
-    libc.src.__support.ctype_utils
 )
 
 add_entrypoint_object(
@@ -474,102 +424,6 @@ add_entrypoint_object(
     .string_utils
     .memory_utils.inline_memset
 )
-
-# Helper to define a function with multiple implementations
-# - Computes flags to satisfy required/rejected features and arch,
-# - Declares an entry point,
-# - Attach the REQUIRE_CPU_FEATURES property to the target,
-# - Add the fully qualified target to `${name}_implementations` global property for tests.
-function(add_implementation name impl_name)
-  cmake_parse_arguments(
-    "ADD_IMPL"
-    "" # Optional arguments
-    "" # Single value arguments
-    "REQUIRE;SRCS;HDRS;DEPENDS;COMPILE_OPTIONS;MLLVM_COMPILE_OPTIONS" # Multi value arguments
-    ${ARGN})
-
-  if("${CMAKE_CXX_COMPILER_ID}" MATCHES "Clang")
-    # Note that '-mllvm' needs to be prefixed with 'SHELL:' to prevent CMake flag deduplication.
-    foreach(opt IN LISTS ADD_IMPL_MLLVM_COMPILE_OPTIONS)
-      list(APPEND ADD_IMPL_COMPILE_OPTIONS "SHELL:-mllvm ${opt}")
-    endforeach()
-  endif()
-
-  if("${CMAKE_CXX_COMPILER_ID}" MATCHES "GNU")
-    # Prevent warning when passing x86 SIMD types as template arguments.
-    # e.g. "warning: ignoring attributes on template argument ‘__m128i’ [-Wignored-attributes]"
-    list(APPEND ADD_IMPL_COMPILE_OPTIONS "-Wno-ignored-attributes")
-  endif()
-
-  add_entrypoint_object(${impl_name}
-    NAME ${name}
-    SRCS ${ADD_IMPL_SRCS}
-    HDRS ${ADD_IMPL_HDRS}
-    DEPENDS ${ADD_IMPL_DEPENDS}
-    COMPILE_OPTIONS ${libc_opt_high_flag} ${ADD_IMPL_COMPILE_OPTIONS}
-  )
-  get_fq_target_name(${impl_name} fq_target_name)
-  set_target_properties(${fq_target_name} PROPERTIES REQUIRE_CPU_FEATURES "${ADD_IMPL_REQUIRE}")
-  set_property(GLOBAL APPEND PROPERTY "${name}_implementations" "${fq_target_name}")
-endfunction()
-
-# ------------------------------------------------------------------------------
-# bcmp
-# ------------------------------------------------------------------------------
-
-function(add_bcmp bcmp_name)
-  add_implementation(bcmp ${bcmp_name}
-    SRCS ${LIBC_SOURCE_DIR}/src/string/bcmp.cpp
-    HDRS ${LIBC_SOURCE_DIR}/src/string/bcmp.h
-    DEPENDS
-      .memory_utils.memory_utils
-      libc.include.string
-    ${ARGN}
-  )
-endfunction()
-
-if(${LIBC_TARGET_ARCHITECTURE_IS_X86})
-  add_bcmp(bcmp_x86_64_opt_sse2   COMPILE_OPTIONS -march=k8             REQUIRE SSE2)
-  add_bcmp(bcmp_x86_64_opt_sse4   COMPILE_OPTIONS -march=nehalem        REQUIRE SSE4_2)
-  add_bcmp(bcmp_x86_64_opt_avx2   COMPILE_OPTIONS -march=haswell        REQUIRE AVX2)
-  add_bcmp(bcmp_x86_64_opt_avx512 COMPILE_OPTIONS -march=skylake-avx512 REQUIRE AVX512BW)
-  add_bcmp(bcmp_opt_host          COMPILE_OPTIONS ${LIBC_COMPILE_OPTIONS_NATIVE})
-  add_bcmp(bcmp)
-elseif(LIBC_TARGET_OS_IS_GPU)
-  add_bcmp(bcmp)
-else()
-  add_bcmp(bcmp_opt_host          COMPILE_OPTIONS ${LIBC_COMPILE_OPTIONS_NATIVE})
-  add_bcmp(bcmp)
-endif()
-
-# ------------------------------------------------------------------------------
-# bzero
-# ------------------------------------------------------------------------------
-
-function(add_bzero bzero_name)
-  add_implementation(bzero ${bzero_name}
-    SRCS ${LIBC_SOURCE_DIR}/src/string/bzero.cpp
-    HDRS ${LIBC_SOURCE_DIR}/src/string/bzero.h
-    DEPENDS
-      .memory_utils.inline_memset
-      libc.include.string
-    ${ARGN}
-  )
-endfunction()
-
-if(${LIBC_TARGET_ARCHITECTURE_IS_X86})
-  add_bzero(bzero_x86_64_opt_sse2   COMPILE_OPTIONS -march=k8             REQUIRE SSE2)
-  add_bzero(bzero_x86_64_opt_sse4   COMPILE_OPTIONS -march=nehalem        REQUIRE SSE4_2)
-  add_bzero(bzero_x86_64_opt_avx2   COMPILE_OPTIONS -march=haswell        REQUIRE AVX2)
-  add_bzero(bzero_x86_64_opt_avx512 COMPILE_OPTIONS -march=skylake-avx512 REQUIRE AVX512F)
-  add_bzero(bzero_opt_host          COMPILE_OPTIONS ${LIBC_COMPILE_OPTIONS_NATIVE})
-  add_bzero(bzero)
-elseif(LIBC_TARGET_OS_IS_GPU)
-  add_bzero(bzero)
-else()
-  add_bzero(bzero_opt_host          COMPILE_OPTIONS ${LIBC_COMPILE_OPTIONS_NATIVE})
-  add_bzero(bzero)
-endif()
 
 # ------------------------------------------------------------------------------
 # memcmp
