+++ libc/src/strings/bcmp.h
@@ -1,4 +1,4 @@
-//===-- Implementation header for bcopy -------------------------*- C++ -*-===//
+//===-- Implementation header for bzero -------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,16 +6,16 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_STRING_BCOPY_H
-#define LLVM_LIBC_SRC_STRING_BCOPY_H
+#ifndef LLVM_LIBC_SRC_STRINGS_BCMP_H
+#define LLVM_LIBC_SRC_STRINGS_BCMP_H
 
 #include "src/__support/macros/config.h"
 #include <stddef.h> // size_t
 
 namespace LIBC_NAMESPACE_DECL {
 
-void bcopy(const void *src, void *dest, size_t count);
+int bcmp(const void *lhs, const void *rhs, size_t count);
 
 } // namespace LIBC_NAMESPACE_DECL
 
-#endif // LLVM_LIBC_SRC_STRING_BCOPY_H
+#endif // LLVM_LIBC_SRC_STRINGS_BCMP_H
