+++ libc/src/strings/bcopy.cpp
@@ -1,4 +1,4 @@
-//===-- Implementation header for rindex ------------------------*- C++ -*-===//
+//===-- Implementation of bcopy -------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,15 +6,15 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_STRING_RINDEX_H
-#define LLVM_LIBC_SRC_STRING_RINDEX_H
-
+#include "src/strings/bcopy.h"
+#include "src/__support/common.h"
 #include "src/__support/macros/config.h"
+#include "src/string/memory_utils/inline_memmove.h"
 
 namespace LIBC_NAMESPACE_DECL {
 
-char *rindex(const char *src, int c);
+LLVM_LIBC_FUNCTION(void, bcopy, (const void *src, void *dst, size_t count)) {
+  return inline_memmove(dst, src, count);
+}
 
 } // namespace LIBC_NAMESPACE_DECL
-
-#endif // LLVM_LIBC_SRC_STRING_RINDEX_H
