+++ libc/src/strings/index.cpp
@@ -1,4 +1,4 @@
-//===-- Implementation header for strcasecmp --------------------*- C++ -*-===//
+//===-- Implementation of index -------------------------------------------===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,15 +6,16 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_STRING_STRCASECMP_H
-#define LLVM_LIBC_SRC_STRING_STRCASECMP_H
+#include "src/strings/index.h"
 
+#include "src/__support/common.h"
 #include "src/__support/macros/config.h"
+#include "src/string/string_utils.h"
 
 namespace LIBC_NAMESPACE_DECL {
 
-int strcasecmp(const char *left, const char *right);
+LLVM_LIBC_FUNCTION(char *, index, (const char *src, int c)) {
+  return internal::strchr_implementation(src, c);
+}
 
 } // namespace LIBC_NAMESPACE_DECL
-
-#endif // LLVM_LIBC_SRC_STRING_STRCASECMP_H
