+++ libc/src/strings/rindex.h
@@ -1,4 +1,4 @@
-//===-- Implementation header for bzero -------------------------*- C++ -*-===//
+//===-- Implementation header for rindex ------------------------*- C++ -*-===//
 //
 // Part of the LLVM Project, under the Apache License v2.0 with LLVM Exceptions.
 // See https://llvm.org/LICENSE.txt for license information.
@@ -6,16 +6,15 @@
 //
 //===----------------------------------------------------------------------===//
 
-#ifndef LLVM_LIBC_SRC_STRING_BZERO_H
-#define LLVM_LIBC_SRC_STRING_BZERO_H
+#ifndef LLVM_LIBC_SRC_STRINGS_RINDEX_H
+#define LLVM_LIBC_SRC_STRINGS_RINDEX_H
 
 #include "src/__support/macros/config.h"
-#include <stddef.h> // size_t
 
 namespace LIBC_NAMESPACE_DECL {
 
-void bzero(void *ptr, size_t count);
+char *rindex(const char *src, int c);
 
 } // namespace LIBC_NAMESPACE_DECL
 
-#endif // LLVM_LIBC_SRC_STRING_BZERO_H
+#endif // LLVM_LIBC_SRC_STRINGS_RINDEX_H
