+++ libc/test/src/CMakeLists.txt
@@ -58,6 +58,7 @@ add_subdirectory(stdfix)
 add_subdirectory(stdio)
 add_subdirectory(stdlib)
 add_subdirectory(string)
+add_subdirectory(strings)
 add_subdirectory(wchar)
 
 # Depends on utilities in stdlib
