+++ libc/test/src/strings/index_test.cpp
