+++ libc/test/src/strings/rindex_test.cpp
